// src/types/apiData/invoice.ts
export type invoice = {
  id: number;
  booking_id: number;
  user_id: number;
  invoice_number: string;
  invoice_starting_number: string;
  invoice_number_suffix: string;
  company_information: string;
  service_details: string;
  tax_details: string;
  note: string;
  total_amount: number;
  status: string;
  created_by: string;
  deleted_by: string;
  is_deleted: boolean;
  deleted_at: string;
  created_at: string;
  updated_at: string;
  user: {
    id: number;
    first_name: string;
    last_name: string;
    email: string;
  };
  booking: {
    id: number;
    booking_number: string;
    shortlet_id: number;
    user_id: number;
    check_in_date: string;
    check_out_date: string;
    check_in_time: string;
    check_out_time: string;
    number_of_days: number;
    number_of_guests: number;
    guest_breakdown: string;
    currency: string;
    caution_fee: number;
    tax_fee: number;
    base_price: number;
    discount_fee: number;
    total_price: number;
    exchange_rate: string;
    caution_fee_NGN: number;
    tax_fee_NGN: number;
    base_price_NGN: number;
    discount_amount_NGN: number;
    total_price_NGN: number;
    status: string;
    payment_status: string;
    payment_method: string;
    promo_code: string;
    notes: string;
    channel: string;
    payment_expires_at: string;
    is_deleted: 1 | 0;
    deleted_at: string;
    created_at: string;
    updated_at: string;
    transferred_from: string;
    transferred_to: string;
    visitor_code: string;
    account_name: string;
    account_number: string;
    bank_name: string;
    has_requested_refund: 0 | 1;
    shortlet: {
      id: number;
      name: string;
      slug: string;
      description: string;
      location: string;
      currency: string;
      price: number;
      caution_fee: number;
      tax_fee: number;
      no_of_bedrooms: number;
      no_of_bathrooms: number;
      min_guests: number;
      max_guests: number;
      point_of_interest: string;
      cancellation_policy: string;
      availability_status: string;
      is_deleted: number;
      deleted_at: number;
      created_at: string;
      updated_at: string;
      room_option_id: number;
      city: string;
      state: string;
      country: string;
      longitude: string;
      latitude: string;
      location_group_id: string;
    };
  };
};